Create improved.c
Login functionality: Once a user has registered an account, they should be able to log in to their account securely. You will need to add a login option to the main menu and prompt the user to enter their phone number and password. You will then need to read the data from the file and compare the entered phone number and password with the data in the file. If the login credentials match, allow the user to access their account.

Account summary: Once a user has logged in, they should be able to view their account summary, including their account number, phone number, balance, and transaction history. You will need to read the data from the file and display it in a user-friendly format.

Deposit functionality: Users should be able to deposit money into their account. You will need to add a deposit option to the main menu and prompt the user to enter the amount they wish to deposit. You will then need to update the balance in the file and display the new balance to the user.

Withdrawal functionality: Users should be able to withdraw money from their account. You will need to add a withdrawal option to the main menu and prompt the user to enter the amount they wish to withdraw. You will then need to check if the user has sufficient balance and update the balance in the file accordingly.

Transfer functionality: Users should be able to transfer money to another account. You will need to add a transfer option to the main menu and prompt the user to enter the recipient's phone number and the amount they wish to transfer. You will then need to check if the recipient's account exists and if the user has sufficient balance to make the transfer. You will also need to update the balances in both accounts and display a confirmation message to the user.

Error handling: You will need to handle errors gracefully and provide informative error messages to the user in case of incorrect inputs or other errors.

Security: You will need to ensure the application is secure, for example, by hashing the passwords before storing them in the file, or using SSL/TLS to encrypt communication between the user's browser and the server.